// trova-job/src/app/models/abilities.model.ts
export class UserAbilities extends Object {
  canReadUsers?: boolean;
  canCreateUsers?: boolean;
  canUpdateUsers?: boolean;
  canDeleteteUsers?: boolean;
  canReadJobs?: boolean;
  canCreateJobs?: boolean;
  canUpdateJobs?: boolean;
  canDeleteJobs?: boolean;
  canReadJob?: boolean;
  canCreateJob?: boolean;
  canUpdateJob?: boolean;
  canDeleteJob?: boolean;
  canUpdateProfile?: boolean;
}